refactor: modularize server architecture with separate auth and oRPC handlers
