+++ apps/backend/package.json
@@ -34,6 +34,8 @@
 		"@types/node": "^22.18.12",
 		"@types/pg": "^8.15.5",
 		"dotenv": "^17.2.3",
+		"pino-http": "^11.0.0",
+		"pino-pretty": "^13.1.3",
 		"tsx": "^4.19.2"
 	}
 }
